requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
Pillow==11.2.1
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
pandas==2.3.0
numpy==2.2.6
opencv-python-headless==4.11.0.86
pdf2image==1.17.0
pdf2docx==0.5.8
pdf2pptx==1.0.5
pdfplumber==0.11.7
PyMuPDF==1.20.1
pypdfium2==4.30.1
lxml==6.0.0
fpdf==1.7.2
docx2pdf==0.1.8
pyheif==0.8.0
pillow-heif==1.0.0
